Added new applicative-style infix combinators (`describedBy`, `on`, `sample`) and made `propertyN` to use them.
Also, the notion of `Property` changed: It is now an alias for `PropertyBuilder Bool` (necessary change to enable the aforementioned combinators), which should be transparent anyway. A property is characterized by its `wrappedGenerator`, which is a `Generator` of `TestResult`s. Optionally, `requestedSamples` specifies the number of times to sample `TestResult`s from the `Generator`.

Actual sampling and evaluation takes place in `check`: It produces a `TestOutput` by sampling the passed properties.